♻️  Refactor types and interfaces
