+++ src/steps/configureScripts.ts
@@ -1,6 +1,6 @@
 import editJson from 'edit-json-file';
-import { ExtraModulesAnswer } from '../models/Answers';
-import { GeneralAnswers } from '../models/answerChoice';
+import { ExtraModulesAnswer } from '../models/ChoiceAnswers';
+import { GeneralAnswers } from '../models/PromptAnswers';
 
 function configureScripts(editablePackageJson: editJson.JsonEditor, answers: GeneralAnswers): void {
   const prod = answers.extras.includes(ExtraModulesAnswer.Crossenv) ? 'cross-env NODE_ENV=production' : '';
