+++ src/steps/initNpm.ts
@@ -1,7 +1,7 @@
 import path from 'path';
 import editJson from 'edit-json-file';
-import { GeneralAnswers } from '../models/answerChoice';
-import { Paths } from '../models/paths';
+import { Paths } from '../models/Paths';
+import { GeneralAnswers } from '../models/PromptAnswers';
 import exec from '../utils/exec';
