+++ src/steps/installOtherDependencies.ts
@@ -2,9 +2,9 @@ import { promises as fs } from 'fs';
 import path from 'path';
 
 import FilesData from '../FilesData';
-import { ExtraModulesAnswer } from '../models/Answers';
-import { GeneralAnswers } from '../models/answerChoice';
-import { Paths } from '../models/paths';
+import { ExtraModulesAnswer } from '../models/ChoiceAnswers';
+import { Paths } from '../models/Paths';
+import { GeneralAnswers } from '../models/PromptAnswers';
 import exec from '../utils/exec';
